Add Gradle module metadata to Bintray uploads
__Motivation__

An issue in the current version of the Bintray plugin prevents it from uploading the Gradle module metadata file.

__Modification__

Add the metadata files to the bintray task config.

Note that a more involved workaround is proposed in [1], this PR sends a workaround that is purely file based.

__Results__

Gradle module metadata are published.

[1] https://github.com/bintray/gradle-bintray-plugin/issues/229