Fix Bintray upload of Gradle module metadata
__Motivation__

The recently introduced fix for Bintray plugin's Gradle module metadata issue is not working because the Bintray plugin copy task (`_bintrayRecordingCopy`) runs before the generation of the Gradle metadata (`generateMetadataFileForMavenJavaPublication`).

This was not detected locally because the Gradle metadata was preexisting in an unclean build when `bintrayUpload` was tested.

__Modifications__

- Use the workaround suggested in [1],
- Review application of `maven-publish` plugin as it now must be done before Bintray.

__Results__

Gradle module metadata files are uploaded to Bintray (hopefully 😅).

[1] https://github.com/bintray/gradle-bintray-plugin/issues/229