Adding a ProtobufSerializationProvider to do generic protobuf serialization for any Message type that it can see a Parser for.
Motivation:

A common protobuf serialization implementation would be benefitial for most all services that use protobufs as a data encoding.

Modifications:

- Added servicetalk-data-protobuf package with a SerializationProvider and accompanying tests.

Result:

Any project needing ServiceTalk Buffer to Protobuf serialization/deserialization should be able to bring this project in and use it directly with little to no configuration or customization.