Prevent conflicts between gRPC generated type names
__Motivation__

If a proto file does not configure the `java_outer_classname` option, codegen will try to infer a name for this outer class name based on the proto file name. This creates conflicts if the proto file name leads to the same name as one of the generated class.

For example, consider `account_service.proto` which contains `service Account` and does not configure `java_outer_classname`. Codegen will then output an outer class named `AccountService` which will contain a service interface named `AccountService`, which does not compile.

__Modifications__

- Add tests that create conflict between generated class names
- Deconflict generated class names

__Results__

Generated gRPC code is less likely to be wrong.