Remove `ConnectionContext` argument from `ChannelInitializer.init`
Motivation:

None of our current `ChannelInitializer` implementations use
`ConnectionContext` argument of `init` method. Because this is internal
API it's better to simplify it. Removing requirement for
`ConnectionContext` argument also helps to avoid unnecessary allocation
of dummy `ConnectionContext` for TCP `ChannelInitializer` with ALPN.

Modifications:

- Remove `ConnectionContext` argument from `ChannelInitializer.init`;
- Update all places where we use `ChannelInitializer` to account for
this change;
- Remove `AlpnConnectionContext`;

Result:

Simplified API of `ChannelInitializer`.