Fix cross-components links in documentation website
Motivation:

1. When a `{page-version}` is not specified for a cross-component
link Antora links to the latest version of that component.
We need to specify a version for each cross-component link.
2. When the page from a different component is opened, the whole
navigation bar changes and show only links of that component.

Modifications:

- Create `nav-versioned.adoc` for each component that specifies a full
link for each page;
- Include `nav-versioned.adoc` from each component into one navigation
file of the main `servicetalk` component;
- Each component uses the aggregated `nav-versioned.adoc` from the main
`servicetalk` component to keep the navigation bar consistent;
- Specify a full `xref` for each link, including a component ID and
the version number;

Result:
1. Correct links across different docs versions and components.
2. Consistent navigation bar for all components.