Fix broken links when you open docs on GitHub (#809)
Motivation:

We use `{source-root}` attribute as a path to the source
code when we generate the docs website. However, GitHub
does not understand it. Therefore, all links are broken
when users open `*.adoc` files on GitHub.

Modifications:

- Use `env-github` to understand that an `*.adoc` file
is opened on GitHub;
- Use `page-origin-refname` instead of manual `git-ref`;

Result:

All links to source code work in both cases: on GitHub
and on the docs website.