Flush optimization for aggregated API on server are broken
__Motivation__

If a response is safe to be aggregated; while writing to the connection, we update the flush strategy for the connection appropriately (use flush-on-end) and revert the update when that response in written. However, the trigger point for revert is incorrect as it is composed on the flattened response `Publisher`. When the components of a response is emitted from the `Publisher` they may be queued in the offloading `Executor` or on the netty event loop, so reverting the strategy when the response `Publisher` terminates may revert it before the response is written on the channel. This leads to falling back to the flush-on-each strategy.

__Modification__

The appropriate place to infer the trigger points for reverting the flush-strategy update is in the callbacks to `WriteEventsListener` as they are invoked strictly in line with the writes on the channel and hence all actions are guaranteed to be ordered. However, `WriteEventsListener#itemWritten()` does not provide the item which is written and hence it is difficult to revert the strategy from that method. `WriteEventListener#itemWritten()` now also provides the item that was written. Now, implementation can delineate write boundaries by looking at the item that was written without having to add hooks to the original `Publisher` that was written.

Added `SplittingFlushStrategy` that can delineate writes from a single stream into logical units which can then be controlled by different `FlushStrategy` instances. HTTP server uses this implementation in place of the original more complex `CompositeFlushStrategy` implementation.

Also, added tests with `EmbeddedChannel` to deterministically inspect flush boundaries and verify the expected behavior.

__Result__