Reconsider API for configuring different protocols (#816)
Motivation:

Existing approach of configuring HTTP protocol versions and the version-
specific options are not straight forward and requires users to configure
TLS ALPN separately. It complicates the bootstrap of the application and
users can easily misconfigure the client or server. The version-specific
options are mixed together and it's hard to distinguish them.

Modifications:

- Introduce `HttpProtocolConfig` interface to initiate HTTP config;
- Add two version-specific configuration interfaces:
`H1ProtocolConfig` and `H2ProtocolConfig`;
- Move all HTTP/1.1 related configuration options from builders to the
`H1ProtocolConfig`;
- Move all HTTP/2 related configuration options from builders to the
`H2ProtocolConfig`;
- Add `H1ProtocolConfigBuilder`, `H2ProtocolConfigBuilder`, and
`HttpProtocolConfigs` that creates these builders and keeps the default
configurations;
- Remove all ALPN-related configs from `SecurityConfigurator` API;
- Automatically configure ALPN based on the provided protocols when TLS
is enabled;
- Throw an exception if ALPN is not supported in the current environment
at the build time;
- Refactor internal TCP and HTTP config classes;
- Use `HttpProtocolConfig` for configuring grpc;
- Update tests;

Result:

Users can provide a list of required HTTP protocols and configure them
separately. ALPN will be configured automatically based on the list of
protocols when TLS is enabled.