Always run `buildExecutable` task in `servicetalk-grpc-protoc` (#823)
Motivation:

We depend on the results of `buildExecutable` in examples and grpc-netty
modules. Because this is a custom task we have to include it in a task
dependency graph. Instead of adding `buildExecutable` as a dependent
task for each other task, we should always include it when classes are
generated.

Modifications:

- Use `finalizedBy` instead of `dependsOn` to add `buildExecutable` to
the tasks dependency graph;

Result:

`buildExecutable` will always be executed when we generate classes of
`servicetalk-grpc-protoc` module.