Use `AddressUtils.localAddress` in tests where it was missed
Motivation:

We should never start a server for all interfaces (0.0.0.0) in tests,
because in some environments additional permissions may be required.
Servers should use loopback address only in tests and we provide a
utility that creates it with random port.

Modifications:

- Use `AddressUtils.localAddress` in tests where it was missed;

Result:

Servers in tests do not start using `0.0.0.0` interface.