Support for Single.fromCallable(), Single.fromSupplier(), Completable.fromRunnable()
Motivation:

Currently, if we want to take a piece of non-asynchronous code and create a Single out of it for a response, we’d have to do one of the following:

* Use CompletableFuture or a Mono to encapsulate the code, and then use Single.fromStage()
* Use Single.defer(() -> Single.succeeded(…))

While this works, it’s an extra step that can be provided as a shortcut in the Single class itself.  Both Mono (Project Reactor) and CompletableFuture offer ways of creating deferred executions via a Supplier, Callable, or Runnable.  However, to use the streaming server it has to be of type Single.  It makes sense for Single to have these capabilities directly.

Modifications:

- Added fromCallable() and fromSupplier() to Single
- Added fromRunnable() to Completable
- Added related tests

Result:

Increased simplicity