Parse headers with no whitespaces (#856)
Motivation:

The whitespace between the `:` and a header value is optional [1] but we
consider it as required [2] hence skip a character after the `:`. As the
result, header value may miss the first character.

[1] https://tools.ietf.org/html/rfc7230#section-3.2
[2] `header-field = field-name ":" OWS field-value OWS`

Modifications:

- Do not skip a character after `:`;
- Add more tests for headers with no whitespace and for mixed
whitespaces;

Result:

Headers parser does not skip the first value character when there is no
whitespace after `:`.