Introduce automatic retries
__Motivation__

There are a few expected scenarios where ServiceTalk may use a connection for making request when the connection is already closed. This happens because there is an inherent race between a peer closing a connection and the load balancer choosing a connection. Due to this race, load balancer may end up picking a connection which is already closed but the close notification has not reached the load balancer. In such scenarios, when load balancer chooses an unusable connection writing a request will fail and the users are expected to retry these exceptions. Since in these cases we will never write the request on the wire, these requests are always safe to retry automatically.

__Modification__

We already do a form of automatic retry which is when the load balancer does not have any usable hosts. Since, this new addition is just another form of automatic retry, instead of adding a top-level builder config, now introduced an abstraction called `AutomaticRetryStrategyProvider` which provides a retry strategy for automatic retries. The client builders then can disable or modify this strategy if required.

__Result__

ServiceTalk now automatically retry expected exceptions which are safe to retry.