Enhance SerializationException message from `http-api` (apple#878)
Motivation:

When users attempt to deserialize a response that has the wrong
content-type, they receive an exception that lists all the headers,
but with <filtered> as the value, making it hard to diagnose.

Modifications:

- Add a new debug filter `DEFAULT_DEBUG_HEADER_FILTER`
to the class `HeaderUtils`.
- Modify the following classes [`DefaultClassHttpDeserializer`,
`DefaultTypeHttpDeserializer`, `FormUrlEncodedHttpDeserializer`,
`HttpStringDeserializer`] and change the method `checkContentType`
to call `headers.toString(DEFAULT_DEBUG_HEADER_FILTER)`.
- add tests to cover the new SerializationException messages.

Result:

Better debugging in SerializationException when attempting
to deserialize invalid 'content-type'.