Clean up the state before and after docs publication
Motivation:

As part of the docs website publication we add additional
git worktree and branch. After the publication is complete we
clean up this state. However, if the script failed with an
error, the second run will start from the intermediate git state.

Modifications:

- wrap clean up steps to a function
- invoke the clea up function before and after the docs publication;

Result:

We can safely re-run `publish-docs.sh` script even if the previous
run failed.