gRPC-server should return `UNIMPLEMENTED` when route is not found
Motivation:

When gRPC-server doesn't have a requested route it should respond
with an error status `UNIMPLEMENTED` to propagate error to the
client.

Modifications:

- Change `NOT_FOUND_SERVICE` to return `UNIMPLEMENTED` grpc-status;
- Add a test to verify that client throws `GrpcStatusException`
when it requests a route that is not registered on the server;

Result:

Client throws `GrpcStatusException` when it requests a route that
is not registered on the server.