H2 client ignores query parameters (#933)
__Motivation__

When using H2 as the transport with HTTP clients, if query parameters are present in the request, they get ignored while writing the request on the network.

__Modification__

`H2ToStH1ClientDuplexHandler` is using `request.path()` as the value for the `:path` pseudo-header in H2. `path()` omits query parameters. Now using `requestTarget()`.

__Results__

Query parameters are preserved when using H2 transport.