gRPC `BlockingResponseStreamingRoute` expects only one request item
Motivation:

gRPC `BlockingResponseStreamingRoute` is implemented as a conversion
based on `BlockingStreamingRoute`, but must ensure that there is
only one proto item in the request and fail if there are more.
Otherwise, it does not drain request `BlockingIterable` and may lead
to the leaked resources.

Modifications:

- Throw if `BlockingResponseStreamingRoute` sees the second item
in the request `BlockingIterable`;
- Require non-null response item in `BlockingRequestStreamingCall`;
- Add tests to verify these cases work as expected;

Result:

`BlockingResponseStreamingRoute` fails if a client sends more than
one proto items in the request.