WriteStreamSubscriber and WriteListener cleanup (#923)
Motivation:
The WriteListener interface has grown over time and its naming conventions have
become less clear. For example "closeGracefully" implies that some action is
expected to be taken to account for pending writes, but in reality this is just
a notification that the transport's outbound side has been closed and should be
interpreted accordingly by the implementation.
The CloseHandler introduced a new class and public static variable which is
intended for use in the http/2.0 codec. However the CloseHandler isn't protocol
specific and the naming of this class implies an inverted dependency
relationship.

Modifications:
- Rename WritableListener to ChannelOutboundListener. Rename method names on
this interface to indicate they are notification of channel events and
implementations can react however they need to.
- Rename H2ProtocolHandler to ProtocolOutboundCloseEventHandler to be more
agnositc to specific protocols which may use this class and associated public
static singleton instance.
- DefaultNettyConnectionTest writes a static buffer, but doesn't duplicate() on
each use. The test now duplicates the buffer to avoid shared indexes.
- Avoid forcing error logs from WriteSingleSubscriber because the async
control flow has already been completed and the additional noise/overhead of the
log may not be necessary.

Result:
More clarity around ChannelOutboundListener interfaces and implementations.