Don't re-use *Exceptions when stored in static fields if enableSuppre… (#958)
Motivation:

Have a static instance of an Exception which does not set enableSuppression to false when constructed is considered to be dangerous as it may escape to the user where an user may add more exceptions to it. The problem with this is that the internal storage will grow and grow over time and never be GC'ed. This can lead to OOME situations.

Modifications:

Create new *Exceptions when needed and remove usage of static stored *Exceptions if needed.

Result:

No more risk of OOME due user calling `addSuppressed` on exceptions that are propagated through the code


Co-authored-by: Nitesh Kant <nitesh_kant@apple.com>