Ensure `TcpConnector#connect` does not emit error after success (#967)
__Motivation__

`TcpConnector#connect` eagerly terminates the `Single` when the channel is registered. It also terminates the `Single` with failure if connect fails.
This violates the contract of the `Subscriber` that there should only be a single termination signal.

__Modification__

As we have to complete the returned `Single` with the channel as soon as the channel is registered to ensure we do not miss channel events and we also have to propagate connect failures, we can not abide by the `Single<Channel>` contract. Instead of returning `Single<Channel>` now returning `Single<Connection>` and do the translation of `Channel` -> `Connection` using a supplied `Function`.

__Result__

`TcpConnector#connect` abides by the `Subscriber` contract.