Initial support for windows development (#977)
Motivation:
ServiceTalk currently doesn't build on windows. This can be limiting for
folks who are comfortable with that development environment.

Modifications:
- grpc-protoc project needs refactoring of the plugin in order to
execute on unix and windows platforms. The distribution is now
consistent with a script as the executable and co-located uber jar
containing the plugin logic.
- ignore/modify tests that failed when running locally on windows, there
will be followup PRs for other tests that sporadically fail.

Result:
ServiceTalk can now be built on windows.