gRPC protoc script pushd not found (#982)
Motivation:
The gRPC protoc unix script is expected to execute in a /bin/sh environment, but it depends upon pushd/popd which may not be available in /bin/sh. This may lead to a 'pushd: not found' error.

Modifications:
- Avoid using pushd in the unix script and manually save/restore the directory

Result:
gRPC protoc script can execute in limited /bin/sh environments without
failure.