gRPC gradle plugin write script to buildDir
Motivation:
servicetalk-grpc-gradle-plugin was writing the executable script files
to the temp directory and not cleaning them up. There was also a bug in
a variable substitution which lead to compile errors.

Modifications:
- Write to the buildDir instead of temp, and don't rewrite the file on
every build if it already exists since the version information is in the
file name.
- Fix the variable naming bug

Result:
No more temp files for servicetalk-grpc-gradle-plugin and one less
variable naming bug.