Replace ConcurrentSubscription with DelayedCancellable when synchronous request(Long.MAX_VALUE)
Motivation:
There are some operators which convert from Publisher to Completable or
Single. Since the demand for Completable and Single is implicit on
subscribe(..) these operators need to simulate the demand in
onSubscribe. In order to prevent concurrent access on the Subscription
with the downstream Subscriber and the synchornous requestN call it is
wrapped in a ConcurrentSubscription. However ConcurrentSubscription has
to coordinate between two methods and is inheritly more complicated than
DelayedCancellable. We can use DelayedCancellable which is a simpler
operation at the cost of requesting demand before propagating
synchronous cancel operation. This isn't perceived as a deal breaker
because demand is implicit anyways, and cancel is best effort.

Modifications:
- When converting from Publisher to Completable or Single use
DelayedCancellable instead of ConcurrentSubscription

Result:
Less complicated and less expensive conversion from Subscription to
Cancellable when converting from Publisher to Single or Completable.