AsyncContextInMemoryScopeManager to restore previous Scope on close()
Motivation:
AsyncContextInMemoryScopeManager currently doesn't restore the previous
Scope when the current Scope is closed. This behavior is provided for
ThreadLocalScope and allows users to use try-with-resources type
constructs to create sub-spans more easily.

Modifications:
- AsyncContextInMemoryScopeManager to save/restore the previous Scope
when a new Scope is activated
- AsyncContextInMemoryScopeManager to remove the currentScope() hack so
lifetime of Scopes is more narrowly/correctly constrainted. This means
that filter ordering is important and use of after* operators should be
done with care!

Result:
AsyncContextInMemoryScopeManager has more consistent behavior with
ThreadLocalScope, and lifetimes of Scopes are more narrowly/correctly
constrainted.