Introduce `HttpLoadBalancerFactory` (#994)
__Motivation__

Today `HttpClientBuilder` takes two separate implementations of `LoadBalancerFactory` as well as `Function` as a protocol binder. As `LoadBalancerFactory` is protocol neutral, it does make sense to have the protocol binding separate but having two separate entities makes the generic ugly and it is hard to understand what the parameters stand for.

__Modification__

Introduce `HttpLoadBalancerFactory` that combines these two aspects and makes the usage easier.
Add a `DefaultHttpLoadBalancerFactory` to provide the default implementation using Round robin load balancer and in the future more load balancer implementations geared towards HTTP.

__Result__

Easier to understand interface for HTTP layer load balancing.