Change when* operators to before* instead of after* (#1005)
Motivation:
The when* operators provide call backs to take action when that action
isn't sensitive to execution order. The current implementation of the
when* operators uses after* semantics, despite the order being explicitly
undefiend by the API. The after* semantics may result in user code
executing in unexpected times, espeically when local state is involved.
For example tracing filters utilize AsyncContext and trigger the state
cleanup on the response payload Publisher termination. This filter must
be appended first in order for the trace information to be initialized
correctly and if any subsequent filters apply a payload Publisher
transformation with after* semantics the terminal methods will be
invoked after the tracing Scope has been closed. If subsequent operators
use before* semantics they will not be subject to this ordering issue.

Modifications:
- Change Publisher, Single, Completable when* operators to use before*
semantics instead of after* semantics.

Result:
when* operators use before* semantics to reduce the risk that local
state managed outside the scope of the operator flow is cleaned up when
user code executes.