`RoundRobinLoadBalancer` should compose connection closure with error response (#1006)
Motivation:

Dangling subscribes is usually a bad idea. As RRLB is not composing the
subscribe, it will still keep closing the connection after it returns a failure.
Worse if the closure does not complete, it will keep this subscribe around
forever (and so the resources associated with the subscribe).

Modification:

- Compose new connection closure with the `Single.failed` returned from
`flatMap`;
- Add a test to verify that new connection is closed if it's rejected by
selector;

Result:

A new connection is closed before RRLB fails the returned `Single` from
`selectConnection`.