Promote Single.TerminalSignalConsumer to top level interface (#1008)
Motivation:
TerminalSignalConsumer for Completable and Publisher is a top level
type, but Single has an internal type for the same purpose. This makes
it more challenging to expose utilities classes in a consistent manner
and also may lead to naming collisions.

Modifications:
- Move Single.TerminalSignalConsumer to a top level interface and rename
to SingleTerminalSignalConsumer to avoid naming collisions.
- Expose RunnableTerminalSignalConsumer and re-use it in
BeforeFinallyOnHttpResponseOperator.

Result:
More consistent type definition scheme for TerminalSignalConsumer and
SingleTerminalSignalConsumer and ability to expose conversion/utility
methods outside the scope of the Source.