Publisher#flatMapMergeSingle potential queue drain visibility issue
Motivation:
Publisher#flatMapMergeSingle uses the
ConcurrentUtils#drainSingleConsumerQueue to drain the pending signal
queue. However this method relies upon the queue's isEmpty() method
providing visibility as to if an item has been inserted into the queue
since the last poll information. We should avoid leveraging the queue's
visibility guarantees (which may not hold) and instead leverage the
concurrency avoidance for single consumer drain to provide visibility.

Modifications:
- Remove ConcurrentUtils#drainSingleConsumerQueue
- Introduce ConcurrentUtils#acquirePendingLock and releasePendingLock
methods that allow for notification if another thread attempted to
acquire the lock while it was being held. If all threads that publish
and event (e.g. insert into a queue) use this lock it can ensure no
concurrency and visibility.

Result:
No dependency on queue visibility semantics for draining the signal
queue in Publisher#flatMapMergeSingle.