DNS SRV query support (#1018)
Motivation:
Our DNS client currently doesn't support
[SRV Resource Records](https://tools.ietf.org/html/rfc2782) which is a
commonly used type of resolution.

Modifications:
- Rename DnsServiceDiscoverer to DnsClient internally to clarify the
feature set is inteded to cover dns client use cases.
- Remove the ServiceDiscovererFilter related classes as the only
implementations are specific to DNS, can be replicated with simple
Publisher#retryWhen operator, and the general utility of these broad
APIs across different clients/protocols is yet to be demonstrated.
- Remove tracking of all outstanding streams and close() cleanup to
reduce state. Subscribers of these sources (e.g. LoadBalancers) are
responsible for cancelling if they are no longer interested in demand.
There is best effort close cleanup when TTLs timeout fire in the event
the DnsClient is closed but the streams are not cancelled (which is not
an expected use case).
- Introduce DNS SRV record support into DefaultDnsClient. Initial
support is limited to a single A/AAAA record resolution per SRV record
due to operator availability (e.g. flatMapMergeSingle) which is
sufficient to get started and can be enhanced later.
- Expose static factory methods for SRV based lookups on HttpClients and
GrpcClients

Result:
DNS supports SRV resolution.